requirements.txt
fastapi
uvicorn
langgraph
qdrant-client
openai
sentence-transformers
aiohttp
python-dotenv
pydantic
pydantic-settings
langchain-openai
langchain-anthropic
tavily-python
langchain-community
redis
sqlmodel
langgraph-checkpoint-sqlite
aiosqlite
